refactor: move polling input functions to utils
