+++ src/plugin/input/src/Input.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+// Resources
+#include "resource/InputManager.hpp"
+
+// Utils
+#include "utils/InputUtils.hpp"
+
+// Plugin
+#include "plugin/PluginInput.hpp"
