+++ src/plugin/input/src/plugin/PluginInput.cpp
@@ -1,0 +1,8 @@
+#include "PluginInput.hpp"
+
+#include "InputManager.hpp"
+
+void ES::Plugin::Input::Plugin::Bind()
+{
+    RegisterResource<ES::Plugin::Input::Resource::InputManager>(ES::Plugin::Input::Resource::InputManager());
+}
