+++ src/plugin/input/src/plugin/PluginInput.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "APlugin.hpp"
+
+namespace ES::Plugin::Input {
+class Plugin : public ES::Engine::APlugin {
+  public:
+    explicit Plugin(ES::Engine::Core &core)
+        : ES::Engine::APlugin(core){
+              // empty
+          };
+    ~Plugin() = default;
+
+    void Bind() final;
+};
+} // namespace ES::Plugin::Input
